+++ .travis.yml
@@ -1,11 +1,15 @@
 language: go
+sudo: false
 
 os:
   - linux
   - windows
   - osx
 go:
-  - 1.11.x
+  - go: 1.11.x
+  - go: 1.12.x
+  - go: 1.13.x
+  - go: 1.14.x
 env:
   - GO111MODULE=on
 install:
