+++ group.go
@@ -6,8 +6,8 @@ package fiber
 
 import (
 	"log"
-	"net/http"
-	"reflect"
+
+	fasthttp "github.com/valyala/fasthttp"
 )
 
 // Group ...
@@ -53,74 +53,64 @@ func (grp *Group) Use(args ...interface{}) *Group {
 		case func(*Ctx):
 			handlers = append(handlers, arg)
 		default:
-			log.Fatalf("Invalid handler: %v", reflect.TypeOf(arg))
+			log.Fatalf("Invalid Use() arguments, must be (prefix, handler) or (handler)")
 		}
 	}
-	path = groupPaths(grp.prefix, path)
-	grp.app.registerMethod("USE", path, handlers...)
+	grp.app.registerMethod("USE", groupPaths(grp.prefix, path), handlers...)
 	return grp
 }
 
 // Connect : https://fiber.wiki/application#http-methods
 func (grp *Group) Connect(path string, handlers ...func(*Ctx)) *Group {
-	path = groupPaths(grp.prefix, path)
-	grp.app.registerMethod(http.MethodConnect, path, handlers...)
+	grp.app.registerMethod(fasthttp.MethodConnect, groupPaths(grp.prefix, path), handlers...)
 	return grp
 }
 
 // Put : https://fiber.wiki/application#http-methods
 func (grp *Group) Put(path string, handlers ...func(*Ctx)) *Group {
-	path = groupPaths(grp.prefix, path)
-	grp.app.registerMethod(http.MethodPut, path, handlers...)
+	grp.app.registerMethod(fasthttp.MethodPut, groupPaths(grp.prefix, path), handlers...)
 	return grp
 }
 
 // Post : https://fiber.wiki/application#http-methods
 func (grp *Group) Post(path string, handlers ...func(*Ctx)) *Group {
-	path = groupPaths(grp.prefix, path)
-	grp.app.registerMethod(http.MethodPost, path, handlers...)
+	grp.app.registerMethod(fasthttp.MethodPost, groupPaths(grp.prefix, path), handlers...)
 	return grp
 }
 
 // Delete : https://fiber.wiki/application#http-methods
 func (grp *Group) Delete(path string, handlers ...func(*Ctx)) *Group {
-	path = groupPaths(grp.prefix, path)
-	grp.app.registerMethod(http.MethodDelete, path, handlers...)
+	grp.app.registerMethod(fasthttp.MethodDelete, groupPaths(grp.prefix, path), handlers...)
 	return grp
 }
 
 // Head : https://fiber.wiki/application#http-methods
 func (grp *Group) Head(path string, handlers ...func(*Ctx)) *Group {
-	path = groupPaths(grp.prefix, path)
-	grp.app.registerMethod(http.MethodHead, path, handlers...)
+	grp.app.registerMethod(fasthttp.MethodHead, groupPaths(grp.prefix, path), handlers...)
 	return grp
 }
 
 // Patch : https://fiber.wiki/application#http-methods
 func (grp *Group) Patch(path string, handlers ...func(*Ctx)) *Group {
-	path = groupPaths(grp.prefix, path)
-	grp.app.registerMethod(http.MethodPatch, path, handlers...)
+	grp.app.registerMethod(fasthttp.MethodPatch, groupPaths(grp.prefix, path), handlers...)
 	return grp
 }
 
 // Options : https://fiber.wiki/application#http-methods
 func (grp *Group) Options(path string, handlers ...func(*Ctx)) *Group {
-	path = groupPaths(grp.prefix, path)
-	grp.app.registerMethod(http.MethodOptions, path, handlers...)
+	grp.app.registerMethod(fasthttp.MethodOptions, groupPaths(grp.prefix, path), handlers...)
 	return grp
 }
 
 // Trace : https://fiber.wiki/application#http-methods
 func (grp *Group) Trace(path string, handlers ...func(*Ctx)) *Group {
-	path = groupPaths(grp.prefix, path)
-	grp.app.registerMethod(http.MethodTrace, path, handlers...)
+	grp.app.registerMethod(fasthttp.MethodTrace, groupPaths(grp.prefix, path), handlers...)
 	return grp
 }
 
 // Get : https://fiber.wiki/application#http-methods
 func (grp *Group) Get(path string, handlers ...func(*Ctx)) *Group {
-	path = groupPaths(grp.prefix, path)
-	grp.app.registerMethod(http.MethodGet, path, handlers...)
+	grp.app.registerMethod(fasthttp.MethodGet, groupPaths(grp.prefix, path), handlers...)
 	return grp
 }
 
@@ -132,23 +122,6 @@ func (grp *Group) Get(path string, handlers ...func(*Ctx)) *Group {
 //
 // https://fiber.wiki/application#http-methods
 func (grp *Group) All(path string, handlers ...func(*Ctx)) *Group {
-	path = groupPaths(grp.prefix, path)
-	grp.app.registerMethod("ALL", path, handlers...)
+	grp.app.registerMethod("ALL", groupPaths(grp.prefix, path), handlers...)
 	return grp
-}
-
-// This function is deprecated since v1.8.2!
-// Please us github.com/gofiber/websocket
-func (grp *Group) WebSocket(path string, handle func(*Ctx)) *Group {
-	log.Println("Warning: app.WebSocket() is deprecated since v1.8.2, please use github.com/gofiber/websocket instead.")
-	path = groupPaths(grp.prefix, path)
-	grp.app.registerWebSocket(http.MethodGet, path, handle)
-	return grp
-}
-
-// This function is deprecated since v1.8.2!
-// Please us github.com/gofiber/recover
-func (grp *Group) Recover(handler func(*Ctx)) {
-	log.Println("Warning: Recover() is deprecated since v1.8.2, please use github.com/gofiber/recover instead.")
-	grp.app.recover = handler
 }
